9:12 PM 13/03/2024
I have partially successfully completed module two. I managed to implement a system that should have worked, checking the balance after every year to see whether it matched or went over the target amount. The problem is that I am receiving an OverflowError so I plan to fix that tomorrow. I imported the math module to try use math.ceil() to round the balance after each year but I realised that it would be better to use round() because I can round to two decimal points to simulate cents.